requirements.txt
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
docopt==0.6.2
Flask==3.1.0
fastapi
uvicorn[standard]
gunicorn==20.1.0
python-multipart
mysql-connector-python
# pin passlib and bcrypt to known-compatible versions
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-jose
PyJWT
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pip==24.3.1
pipreqs==0.4.13
requests==2.32.5
urllib3==2.5.0
Werkzeug==3.1.3
yarg==0.1.10
setuptools<81
aiomysql==0.2.0